SASS-11581 Refactor connector config
